requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.2
opencv-python>=4.4.0
scikit-image>=0.17.2
matplotlib>=3.3.2
Pillow>=8.0.1